要修正を修正済みバージョン

+++ OpenGL_Labyrinth/Frame.cpp
@@ -10,7 +10,7 @@
 Position::Position()
 : x{ 0.0 }, y{ 0.0 } {}
 
-Position::Position(GLfloat xPosi, GLfloat yPosi)
+Position::Position(GLdouble xPosi, GLdouble yPosi)
 : x{xPosi}, y{yPosi} {}
 
 
@@ -18,17 +18,17 @@ Position::Position(GLfloat xPosi, GLfloat yPosi)
 Size::Size()
 :width { 0.0 }, height{ 0.0 } {}
 
-Size::Size(GLfloat width, GLfloat height)
+Size::Size(GLdouble width, GLdouble height)
 : width{width}, height{height} {}
 
 Frame::Frame(Size s, Position p)
 : size(s.width, s.height), position(p.x, p.y) {}
 
-Frame::Frame(Size s, GLfloat x, GLfloat y)
+Frame::Frame(Size s, GLdouble x, GLdouble y)
 : size(s.width, s.height), position(x, y) {}
 
-Frame::Frame(GLfloat width, GLfloat height, Position p)
+Frame::Frame(GLdouble width, GLdouble height, Position p)
 : size(width, height), position(p.x, p.y) {}
 
-Frame::Frame(GLfloat width, GLfloat height, GLfloat x, GLfloat y)
+Frame::Frame(GLdouble width, GLdouble height, GLdouble x, GLdouble y)
 : size(width, height), position(x,y) {}
